Refactor: Added AboutUpActivity to activities package
