+++ app/src/main/java/com/app/venki/up/activities/AboutUPActivity.java
@@ -1,10 +1,13 @@
-package com.app.venki.up;
+package com.app.venki.up.activities;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
+
+import com.app.venki.up.LandingPage;
+import com.app.venki.up.R;
 
 public class AboutUPActivity extends AppCompatActivity {
