+++ app/src/main/res/layout/activity_about_up.xml
@@ -7,7 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".AboutUPActivity"
+    tools:context=".activities.AboutUPActivity"
     android:orientation="vertical">
 
     <TextView
